Implemented a better way of structuring the code
