+++ requirements.txt
@@ -1,1 +1,2 @@
+matplotlib
 numpy==1.24.3
